apps/api/requirements.txt
alembic
fastapi[standard]
python-dotenv
sqlalchemy
psycopg2-binary